requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
greenlet==3.2.4
idna==3.11
lxml==6.0.2
numpy==2.0.2
pandas==2.3.3
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
